Phase 0 WP-F: extract a Tauri-free RecordingService from recording_commands.rs
recording_commands.rs shrinks to thin #[tauri::command] wrappers; all
start/stop/pause/resume orchestration, the fatal-error auto-stop flow, and
the meeting-row/transcript-persistence lifecycle now live in
audio::recording_service as plain async functions taking a RecordingContext
(SharedEventSink + Option<SqlitePool>) instead of an AppHandle.

- The four start_recording* variants collapse into one recording_service::start,
  parameterised on whether device names were given (mirrors the dispatch
  recording_commands used to do).
- Tray refresh moves from scattered update_tray_menu(&app) calls to a
  TrayRefreshingSink that forwards every event and refreshes on
  "recording-state" — the one event every phase transition already emits.
- The fatal-error auto-stop callback (set via RecordingManager::set_error_callback)
  can fire from PipeWire's own thread, not a tokio task; it now spawns onto a
  tokio::runtime::Handle captured up front in start() rather than
  tauri::async_runtime::spawn.
- recording_preferences::load_recording_preferences now takes Option<SqlitePool>
  instead of an AppHandle (it only ever needed the pool); updated its three
  other call sites (import.rs, retranscription.rs).
- transcription::validate_transcription_model_ready and
  speaker_diarization::commands::try_init_for_recording still need an
  AppHandle on this branch (WP-E is converting them concurrently) — isolated
  behind two boxed hooks the command wrapper builds, marked
  TODO(WP-E reconcile).
- Added unit tests for PhaseGuard's release-on-early-return and
  begin_start_phase's refusal while a stop is finalising, on top of the
  existing replay_buffered_segments tests (all moved to recording_service).
